made end-of-screen results GUI human-readable in Grands Prix (before it was just a bunch of numbers, some of which didn't make sense to show) first part of this comnmit accidentally committed in previous commit
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2770 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/gui/grand_prix_ending.cpp
@@ -102,7 +102,7 @@ GrandPrixEnd::GrandPrixEnd()
         // find which kart ended at rank 'i'
         for(unsigned int j=start; j < NUM_KARTS; ++j)
         {
-            if( race_manager->getKartFinalGPRank(j) == i-start )
+            if( (int)race_manager->getKartFinalGPRank(j) == (int)(i-start) )
             {
                 this_kart_id = j;
                 break;
